+++ server/functions/messageReducer.js
@@ -7,6 +7,7 @@ const getUserCurrentQuestion = require("./getUserCurrentquestion");
 const findVerifyStatus = require("./verification/findVerifyStatus");
 const createVerifyStatus = require("./verification/createVerifyStatus");
 const updateVerifyStatus = require("./verification/updateVerifyStatus");
+const messageReducerLogic = require("./MessageReducerLogic");
 const Message = new Messaging();
 const knex = require('./knex')();
 
@@ -16,45 +17,18 @@ MessageReducer = (req,res,next) =>{
   //console.log(message.From);
 
   fetchUser(message.From.substring(1)).then(user=>{
-    //check if there is a verify token for the user
-    //to do so, check who sent the last question, then check the verify table
-    //findVerifyStatus(user.id,).then()
     user = user[0];
     getUserCurrentQuestion(user.id).then(currentQuestion=>{
-      //now we fetch verify status with the returned question's id
       if(currentQuestion)
       {
         findVerifyStatus(user.id,currentQuestion[0].admin).then(result=>{
-          console.log("result is.........................................");
-          console.log(result);
           //now do stuff depending on whether the user verified that admin.
-
           if (result !== null){
             //cast result into array if it isn't one
             if(!Array.isArray(result)) result = [result]
-
             if(result[0].status === "verified"){
               //do stuff
-              if(message.Body.substring(0,2) === "!!" || user.state=== "manage"){
-                //This is the message reducer.
-                let command = message.Body.substring(2).toLowerCase();
-                console.log("Command is " + command);
-                if(command === "skip"){
-                  //drop the most recent question in your account
-                  console.log("you should do the skip logic.");
-                  Message.send("Okay, skipping.", message.From);
-                }else if(command ==="help"){
-                  console.log("show the help logic");
-                  Message.send("Type !!skip to ignore current question, !!update to update your info, and !!help to get this message.", message.From);
-                }else if(command === "revoke"){
-                  Message.send("Okay, revoking.  If you wish to reenable, please send reenable.  Site creator: make sure to let the user retarget this admin later.  Maybe a view revoked.", message.From);
-                  createVerifyStatus(user.id,currentQuestion[0].admin,"revoked");
-                }
-              }
-              else{
-                console.log("That was NOT an account management message, bud.");
-                next();
-              }
+              messageReducerLogic(message, user, currentQuestion[0]);
             }else{
               //check if the user sent reenable
               if(message.Body === "reenable"){
@@ -68,10 +42,6 @@ MessageReducer = (req,res,next) =>{
           }else{
             console.log("not verified, doing the checks for that.")
             //this fires if there is not verified status.  Create verified status based on what the person sent.
-            console.log("user is..........................")
-            console.log(user)
-            console.log("currentQuestion is..........................")
-            console.log(currentQuestion)
             if(message.Body.toLowerCase() === "yes"){
               //user is verified, reply and then create auth token.
               Message.send("Awesome, you are verified!  Here's your question:",message.From);
